% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_ppp}
\alias{calc_ppp}
\title{Calculate Posterior Predictive p-values for meta-analysis results}
\usage{
calc_ppp(mcmc, observed_effects, observed_weights)
}
\arguments{
\item{mcmc}{The data frame containing the Markov Chain Monte Carlo values}

\item{observed_effects}{A vector of observed effect sizes from the
meta-analysis.}

\item{observed_weights}{A vector of observed weights from the
meta-analysis.}
}
\value{
mcmc A vector containing the Postrior Predictive p-values for each of
the discrepancy measures.
}
\description{
This function calculates the Posterior Predictive p-values for the
meta-analysis results
}
