% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diagMCMC}
\alias{diagMCMC}
\title{Perform diagnostics on the Markov Chain Monte Carlo chains}
\usage{
diagMCMC(codaObject, parName = varnames(codaObject)[1])
}
\arguments{
\item{codaObject}{The Markov Chain Monte Carlo chains.}

\item{parName}{A character for the name of the discrepancy measure that is
being tested.}
}
\description{
The function tests for autocorrelation, parameter values, and
density. The results of these tests are saved as jpeg files for easy
reference and insertion into manuscripts.
}
