% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ma_monte_p_value}
\alias{ma_monte_p_value}
\title{Calculate Monte Carlo p-values for meta-analysis results}
\usage{
ma_monte_p_value(simulated, observed_effects, observed_weights, meta_es)
}
\arguments{
\item{simulated}{The data frame of values from the simulated meta-analyses
based on the discrepancy measures.}

\item{observed_effects}{A vector of observed effect sizes from the
meta-analysis.}

\item{observed_weights}{A vector of observed weights from the
meta-analysis.}

\item{meta_es}{The pooled effect size from the meta-analysis.}
}
\value{
simulated a vector containing the Monte Carlo p-values for each of
the discrepancy measures.
}
\description{
This function calculates the Monte Carlo p-values for the
meta-analysis results
}
