% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA-Monte-Carlo.R
\name{monte_carlo_ma}
\alias{monte_carlo_ma}
\title{Monte Carlo Resampling for Meta-Analysis Models}
\usage{
monte_carlo_ma(meta, fixed = FALSE, fileName)
}
\arguments{
\item{meta}{An 'rma' object that contains the results of a meta-analysis
performed in the metafor package.}

\item{fixed}{Logical variable specifying whether the meta-analysis results
were using a fixed-effect model (default) or a random-effects model.}

\item{fileName}{A character variable that serves as the stem for all saved
`   output files.}
}
\value{
monte-p_values A vector containing the Monte Carlo p-values for each
of the discrepancy measures.
}
\description{
This function performs Monte Carlo resampling on meta-analysis
results from the metafor package to assess model fit. This function returns
Monte Carlo p-values for the following discrepancy measures: mean study-level
effects, minimum study-level effects, maximum study-level effects, mean
study-level weights, minimum study-level weights, maximum study-level effects
and pooled effect size. Graphical results are also saved for each of these
discrepancy functions.
}
\examples{
\dontrun{
monte_carlo_ma(meta = re_meta, fixed = FALSE, fileName = "re-meta_monte-carlo")
}

}
