% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_simulation}
\alias{plot_simulation}
\title{Plot the Monte Carlo Resampling results}
\usage{
plot_simulation(simulated, observed_effects, observed_weights, meta_es,
  filename)
}
\arguments{
\item{simulated}{The data frame of values from the simulated meta-analyses
based on the discrepancy measures.}

\item{observed_effects}{A vector of observed effect sizes from the
meta-analysis.}

\item{observed_weights}{A vector of observed weights from the
meta-analysis.}

\item{meta_es}{The pooled effect size from the meta-analysis.}

\item{filename}{A character variable that serves as the stem for all saved
`   output files.}
}
\description{
This function plots the results of the meta-analysis to the
results of the simualted meta-analyses on the seven specified discrepancy
measures.
}
