% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA-PPMC.R
\name{ppmc_ma}
\alias{ppmc_ma}
\title{Posterior Predictive Model Check for Meta-Analysis Models}
\usage{
ppmc_ma(fileName, meta, fixed = FALSE)
}
\arguments{
\item{fileName}{A character variable that serves as the stem for all saved
output files.}

\item{meta}{An 'rma' object that contains the results of a meta-analysis
performed in the metafor package.}

\item{fixed}{Logical variable specifying whether the meta-analysis results
were using a fixed-effect model (default) or a random-effects model.}
}
\value{
ppp A vector containing the PPP-values.
}
\description{
This function performs a posterior predictive model check on
meta-analysis results from the metafor package to assess model fit. This
function returns posterior predictive p-values (PPP-values) for the
following discrepancy measures: mean study-level effects, minimum
study-level effects, maximum study-level effects, mean study-level
weights, minimum study-level weights, maximum study-level effects and
pooled effect size. Graphical results are also saved for each of these
discrepancy functions.
}
\examples{
\dontrun{
ppmc_ma(fileName = "re-meta_ppmc", meta = re_meta, fixed = FALSE)
}
}
