% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate}
\alias{simulate}
\title{Simulate meta-analysis results based on an actual meta-analysis}
\usage{
simulate(effect, effect_sd, weight, weight_sd, studies, iter = 1e+05)
}
\arguments{
\item{effect}{The pooled effect size from the meta-analysis.}

\item{effect_sd}{The standard error of the pooled effect size.}

\item{weight}{The mean value of the vector of weights for the studies
included in the meta-analysis}

\item{weight_sd}{The standard deviation of the vector of weights for the
studies included in the meta-analysis.}

\item{studies}{The number of included studies in the meta-analysis.}

\item{iter}{The number of meta-analyses results to simulate.}
}
\value{
results A data frame containing the value of the discrepancy function
for each of the simulated meta-analyses.
}
\description{
This function simulates results for a specified number of
meta-analyses based on the results from the actual meta-analysis.
}
