% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_ppmc}
\alias{plot_ppmc}
\title{Calculate Posterior Predictive p-values for meta-analysis results}
\usage{
plot_ppmc(mcmc, observed_effects, observed_weights, meta, filename)
}
\arguments{
\item{mcmc}{The data frame containing the Markov Chain Monte Carlo values}

\item{observed_effects}{A vector of observed effect sizes from the
meta-analysis.}

\item{observed_weights}{A vector of observed weights from the
meta-analysis.}

\item{meta}{An 'rma' object that contains the results of a meta-analysis
performed in the metafor package.}

\item{filename}{A character variable that serves as the stem for all saved
`   output files.}
}
\description{
Graph the Posterior Predictive p-values for the meta-analysis
results
}
